+++ .env.example
@@ -22,11 +22,11 @@ NEXTAUTH_SECRET="your-nextauth-secret-change-in-production-must-be-at-least-32-c
 # =====================================================
 # Google Cloud Console에서 발급받은 OAuth 2.0 클라이언트 ID
 # https://console.cloud.google.com/apis/credentials
-GOOGLE_CLIENT_ID="your-google-client-id.apps.googleusercontent.com"
-GOOGLE_CLIENT_SECRET="your-google-client-secret"
+AUTH_GOOGLE_ID="your-google-client-id.apps.googleusercontent.com"
+AUTH_GOOGLE_SECRET="your-google-client-secret"
 
-NAVER_CLIENT_ID="your-navert-client-id"
-NAVER_CLIENT_SECRET="your-naver-client-secret"
+AUTH_NAVER_ID="your-navert-client-id"
+AUTH_NAVER_SECRET="your-naver-client-secret"
 
 # =====================================================
 # 백엔드 API 설정
