+++ src/app/(authenticated)/dashboard/page.tsx
@@ -14,7 +14,7 @@ import { DashboardClient } from "@/components/dashboard/DashboardClient";
 import { StatsCards } from "@/components/dashboard/StatsCards";
 import { WelcomeSection } from "@/components/dashboard/WelcomeSection";
 import { getActionDataOrDefault } from "@/lib/server/action-result-handler";
-import { auth } from "@/lib/server/auth";
+import { auth } from "@/lib/server/auth/auth";
 import { getSelectedFamilyUuid } from "@/lib/server/auth/auth-helpers";
 import { redirect } from "next/navigation";
