--- src/lib/server/auth/index.ts
@@ -1,7 +1,0 @@
-/**
- * 인증 관련 모듈
- *
- * NextAuth 설정 및 인증 유틸리티 함수들을 제공합니다.
- */
-
-export * from "./config";
